Cleaned up test files, added more test for macros, fixed source file rename and move after transfer
