+++ Frends.SFTP.UploadFiles/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.0.1] - 2022-06-07
+### Fixed
+- Cleaned up test files, added more test for macros, fixed source file rename and move after transfer
+
 ## [1.0.0] - 2022-05-24
 ### Added
 - Initial implementation
