+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles.Tests/Lib/Helpers.cs
@@ -33,22 +33,6 @@ namespace Frends.SFTP.UploadFiles.Tests
             return connection;
         }
 
-        internal static void CreateSourceDirectoryWithDate()
-        {
-            var workDir = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../TestData/");
-            var year = DateTime.Now.Year.ToString();
-            var dir = Path.Combine(workDir, "testfolder_" + year);
-            if (!Directory.Exists(dir))
-            {
-                Directory.CreateDirectory(dir);
-                if (Directory.GetFiles(dir).Length == 0)
-                {
-                    File.Copy(Path.Combine(workDir, "SFTPUploadTestFile.txt"), Path.Combine(dir, "SFTPUploadTestFile.txt"));
-                    File.Copy(Path.Combine(workDir, "SFTPUploadTestFile2.txt"), Path.Combine(dir, "SFTPUploadTestFile2.txt"));
-                }
-            } 
-        }
-
         internal static void DeleteSourceDir()
         {
             var year = DateTime.Now.Year.ToString();
