--- Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles.Tests/TestData/testfolder_2022/SFTPUploadTestFile2.txt
@@ -1,1 +1,0 @@
-This is another test file
